requirements.txt
flask
python-dotenv
supabase
requests
torch
pillow
colorthief
clip@git+https://github.com/openai/CLIP.git